+++ src/main/java/com/dinosaur/foodbowl/global/config/security/SecurityConfiguration.java
@@ -1,5 +1,7 @@
 package com.dinosaur.foodbowl.global.config.security;
 
+import com.dinosaur.foodbowl.global.config.security.jwt.JwtAuthenticationFilter;
+import com.dinosaur.foodbowl.global.config.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -25,7 +27,7 @@ public class SecurityConfiguration {
   public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
     http
         .authorizeRequests()
-        .requestMatchers("/docs/*", "/users/sign-up", "/users/sign-in", "/thumbnail/**").permitAll()
+        .requestMatchers("/docs/*", "/sign-up", "/log-in", "/thumbnail/**").permitAll()
         .anyRequest().hasRole("회원")
         .and()
         .httpBasic().disable()
